Rename utils/ulysses.py to utils/sequence_parallel.py
The module's public entry point is the `sequence_parallel` context manager, the counterpart of torch's
`context_parallel`, so the file is named after that rather than after the paper.
