+++ docs/source/concept_guides/sequence_parallelism.md
@@ -253,7 +253,7 @@ The sequence length must be a multiple of `sp_size`. `position_ids` tell attenti
 
 ### Where the all-to-all hooks in
 
-`transformers` models fetch their attention function from `ALL_ATTENTION_FUNCTIONS[config._attn_implementation]` at every forward. `prepare` registers `accelerate.utils.ulysses.UlyssesAttention` under the model's own key, wrapping the original function: it all-to-alls `query`, `key` and `value` from `[batch, heads, local_seq, head_dim]` to `[batch, heads / sp_size, seq, head_dim]`, reads the `position_ids` of the whole sequence (gathered once per forward by a hook on the model, not once per layer), drops the local-shard mask (flash attention reads document boundaries from `position_ids`; `sdpa` gets a block-causal mask rebuilt from them when packed, `is_causal` otherwise), calls the original attention unchanged, and all-to-alls the output back. The wrapper only acts on modules of prepared models, so a reference model in the same process keeps plain attention. The DeepSpeed backend hooks in at the same point.
+`transformers` models fetch their attention function from `ALL_ATTENTION_FUNCTIONS[config._attn_implementation]` at every forward. `prepare` registers `accelerate.utils.sequence_parallel.UlyssesAttention` under the model's own key, wrapping the original function: it all-to-alls `query`, `key` and `value` from `[batch, heads, local_seq, head_dim]` to `[batch, heads / sp_size, seq, head_dim]`, reads the `position_ids` of the whole sequence (gathered once per forward by a hook on the model, not once per layer), drops the local-shard mask (flash attention reads document boundaries from `position_ids`; `sdpa` gets a block-causal mask rebuilt from them when packed, `is_causal` otherwise), calls the original attention unchanged, and all-to-alls the output back. The wrapper only acts on modules of prepared models, so a reference model in the same process keeps plain attention. The DeepSpeed backend hooks in at the same point.
 
 ### Limitations
